Refactor backup script to handle missing files and directories
The backup script has been refactored to handle cases where certain files or directories are missing. Instead of throwing an error, the script now checks if a file or directory exists before attempting to back it up. This prevents unnecessary errors and ensures that only existing files and directories are backed up.

+++ macos/backup.sh
@@ -15,14 +15,25 @@ fi
 
 # Backup the ksh config
 echo "--- Backing up zsh config"
-cp $HOME/.zshrc $BACKUP_DIR/zshrc
-cp $HOME/.zshenv $BACKUP_DIR/zshenv
-cp $HOME/.zprofile $BACKUP_DIR/zprofile
+if [ -f "$HOME/.zshrc" ]; then
+    cp $HOME/.zshrc $BACKUP_DIR/zshrc
+fi
+if [ -f "$HOME/.zshenv" ]; then
+    cp $HOME/.zshenv $BACKUP_DIR/zshenv
+fi
+if [ -f "$HOME/.zprofile" ]; then
+    cp $HOME/.zprofile $BACKUP_DIR/zprofile
+fi
+
 
 # Backup bash
 echo "--- Backing up bash config"
-cp $HOME/.bashrc $BACKUP_DIR/bashrc
-cp $HOME/.bash_profile $BACKUP_DIR/bash_profile
+if [ -f "$HOME/.bashrc" ]; then
+    cp $HOME/.bashrc $BACKUP_DIR/bashrc
+fi
+if [ -f "$HOME/.bash_profile" ]; then
+    cp $HOME/.bash_profile $BACKUP_DIR/bash_profile
+fi
 
 # Backup the aliases
 echo "--- Backing up aliases"
@@ -30,14 +41,30 @@ alias > $BACKUP_DIR/alias
 
 # Backup the VSCode settings
 echo "--- Backing up Visual Studio Code"
-cp -r $HOME/Library/Application\ Support/Code/User $BACKUP_DIR/User
-code --list-extensions > $BACKUP_DIR/vscode-extensions
+if [ -d "$HOME/Library/Application\ Support/Code/User" ]; then
+    mkdir -p $BACKUP_DIR/User
+    cp -r $HOME/Library/Application\ Support/Code/User $BACKUP_DIR/User
+fi
+
+# Backup the VSCode extensions
+# check if code is installed, and if it is, back up the extensions
+if command -v code >/dev/null 2>&1; then
+    echo "--- Backing up Visual Studio Code extensions"
+    code --list-extensions > $BACKUP_DIR/vscode-extensions
+fi
 
 # Backup Terraform
 echo "--- Backing up Terraform"
-cp -r $HOME/.terraformrc $BACKUP_DIR/terraformrc
+if [ -d "$HOME/.terraform.d" ]; then
+    cp -r $HOME/.terraformrc $BACKUP_DIR/terraformrc
+fi
 
 # Backup Git
 echo "--- Backing up Git"
-cp -r $HOME/.gitconfig $BACKUP_DIR/gitconfig
-cp -r $HOME/.gitignore_global $BACKUP_DIR/gitignore_global
+if [ -f "$HOME/.gitconfig" ]; then
+    cp $HOME/.gitconfig $BACKUP_DIR/gitconfig
+fi
+if [ -f "$HOME/.gitignore_global" ]; then
+    cp $HOME/.gitignore_global $BACKUP_DIR/gitignore_global
+fi
+
